pyproject.toml
########################################
# 1) Системные данные для сборки пакета
########################################
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

########################################
# 2) Метаданные проекта и prod-зависимости
########################################
[project]
name = "jarvis"
version = "3.0.0"
description = "Just A Rather Very Intelligent System vol. 3"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
  { name = "Hudrolax", email = "hudro795@gmail.com" }
]

dependencies = [
  "google-api-core==2.25.0",
  "google-api-python-client==2.171.0",
  "google-auth-oauthlib==1.2.2",
  "fastapi==0.115.12",
  "sqlalchemy[asyncio]==2.0.41",
  "alembic==1.16.1",
  "aiohttp==3.11.18",
  "asyncpg==0.30.0",
  "argon2-cffi==25.1.0",
  "PyJWT==2.10.1",
  "uvicorn==0.34.0",
  "pydantic-settings==2.9.1",
  "pgvector==0.4.1",
  "openai==1.91.0",
  "langgraph==0.4.8",
  "langchain==0.3.25",
  "langchain-openai==0.3.24",
  "aiogram==3.20.0.post0"
]

########################################
# 3) dev-зависимости
########################################
[project.optional-dependencies]
dev = [
  "pytest-asyncio==1.0.0",
  "httpx==0.28.1",
  "aiosqlite==0.21.0",
  "mypy==1.16.0",
  "types-passlib==1.7.7.20250602",
  "types-python-jose==3.5.0.20250531",
  "asgi-lifespan==2.1.0",
]

########################################
# 4) Где искать пакеты (каталог app/)
########################################
[tool.setuptools.packages.find]
where = [""]               # корень пайтон-пакетов
exclude = ["tests*"]

########################################
# 7) Линтеры / форматтеры
########################################
[tool.ruff]
line-length    = 120
target-version = "py312"
select         = ["E", "F", "I"]
extend-select  = ["I"]
src            = ["app"]

[tool.black]
line-length    = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2
sections = ["STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
force_sort_within_sections = true
combine_as_imports = false
default_section = "THIRDPARTY"
